feat: Add `RealIP` function
This commit adds a function to return the "real" IP address of a request
by looking for a suitable value in, in order, the `X-Real-Ip` header,
the `X-Forwarded-For` header, and the `RemoteAddr` field of the
`http.Request`. This is useful due to the varying ways and places the
originating IP for a request may live depending on the path it took
through the internet. This function isn't guaranteed to always return a
meaningful value, but it works in most circumstances I've come across.